+++ src/pools/thread/dynamic.ts
@@ -40,7 +40,8 @@ export class DynamicThreadPool<
 
   /** @inheritDoc */
   protected shallCreateDynamicWorker(): boolean {
-    return !this.full && this.internalBusy()
+    return (!this.full && this.internalBusy()) ||
+      (this.minimumNumberOfWorkers === 0 && this.workerNodes.length === 0)
   }
 
   /** @inheritDoc */