+++ tests/pools/thread/dynamic.test.mjs
@@ -179,6 +179,31 @@ Deno.test({
         await pool.destroy()
       },
     )
+
+    await t.step(
+      'Verify that a pool with zero worker works',
+      async () => {
+        const pool = new DynamicThreadPool(
+          0,
+          max,
+          new URL(
+            './../../worker-files/thread/testWorker.mjs',
+            import.meta.url,
+          ),
+        )
+        expect(pool.starting).toBe(false)
+        expect(pool.workerNodes.length).toBe(pool.info.minSize)
+        const maxMultiplier = 10000
+        const promises = new Set()
+        for (let i = 0; i < max * maxMultiplier; i++) {
+          promises.add(pool.execute())
+        }
+        await Promise.all(promises)
+        expect(pool.workerNodes.length).toBe(max)
+        // We need to clean up the resources after our test
+        await pool.destroy()
+      },
+    )
   },
   sanitizeResources: false,
   sanitizeOps: false,